Improve debugging of Constraint Graph
1. Added ConstraintGraph::view() which can be called from
the debugger to view the constraint graph.

2. Added option, -brief-constraint-graph (default true), to
control whether instructions are included in the dump of
the constraint graph nodes (if brief is false). Previously
the brief flag could only be manipulated by editing the
source code and rebuilding.

3. In the non-brief display, use value2String so that
if the value is a function, only the name is given rather
than dumping all of the instructions in the function.

4. In the non-brief display, provide the node Id in any
case, as it takes only a few characters and is likely
to be useful.

5. Correct NormalGepPE::toString(). Due to a presumed
cut and paste error, it was claiming to be a VariantGepPE.