Improve debugging of call graph and value flow graph
1. Added PTACallGraph::view(), callable from debugger,
to display the call graph while debugging.

2. Added VFG::view(), callable from debugger,
to display the [sparse] value flow graph while debugging.

3. Use value2String in a few places to get the special
treatment for functions that it gives (instead of dumping the
IR for the whole function).

4. There were missing cases when dumping a sparse value flow
graph, resulting in assertion violations. I added code to
cover the misssing cases. You may want to check that
the colors and styles I selected are what you want.

5. I modified dumpMSSA() so it unconditionally dumps the
memory SSA without checking for Options::DumpMSSA.
In the sole place where it had been called I put the
Options::DumpMSSA check. This will allow dumpMSSA() to
be called from the debugger if desired without concern
for the setting of Options::DumpMSSA.